MCR-3148 refactor MCRClassificationMappingEventHandler
