+++ mycore-base/src/main/resources/config/mycore.properties
@@ -134,6 +134,8 @@ MCR.Access.AddObjectDefaultRule=true
 
 # Define the Service Flag Handler for objects to set the service flags createdby and modifiedby
   MCR.EventHandler.MCRObject.012.Class=org.mycore.access.MCRAccessCacheEventHandler
+# enable classification mapping as event handler
+  MCR.EventHandler.MCRObject.015.Class=org.mycore.common.events.MCRClassificationMappingEventHandler
 # MCR.EventHandler.MCRObject.017.Class=org.mycore.datamodel.common.MCRServiceFlagEventHandler
 # Define XML store handler
   MCR.EventHandler.MCRObject.020.Class=org.mycore.datamodel.common.MCRXMLMetadataEventHandler
@@ -549,3 +551,9 @@ MCR.URIResolver.ModuleResolver.crypt=org.mycore.crypt.MCRCryptResolver
 
 # comma separated list of allowed prefixes for basket uri
 MCR.Basket.Resolver.AllowList=mcrobject:
+
+##############################################################################
+# XPath Mapping
+##############################################################################
+# Relevant classifications for x-path-mapping
+MCR.Category.XPathMapping.ClassIDs=
