+++ mycore-base/src/test/resources/MCRClassificationMappingEventHandlerTest/dummyClassification.xml
