+++ mycore-base/src/test/resources/MCRClassificationMappingEventHandlerTest/genre.xml
