+++ mycore-base/src/test/resources/MCRClassificationMappingEventHandlerTest/testMcrObject.xml
