+++ mycore-base/src/test/resources/MCRClassificationMappingEventHandlerTest/testMcrObject2.xml
