+++ mycore-oai/src/main/resources/components/oai/config/mycore.properties
@@ -33,8 +33,6 @@ MCR.OAIDataProvider.MetadataFormat.google-culture.Schema=http\://www.manuscript-
 MCR.OAIDataProvider.MetadataFormat.tei.Namespace=http\://www.tei-c.org/ns/1.0
 MCR.OAIDataProvider.MetadataFormat.tei.Schema=http\://www.tei-c.org/release/xml/tei/custom/schema/xsd/tei_all.xsd
 
-# enable classification mapping as event handler
-MCR.EventHandler.MCRObject.015.Class=org.mycore.oai.classmapping.MCRClassificationMappingEventHandler
 # required for deletion policy != 'no'
 MCR.EventHandler.MCRObject.060.Class=org.mycore.datamodel.metadata.history.MCRMetadataHistoryManager
 MCR.EventHandler.MCRDerivate.070.Class=org.mycore.datamodel.metadata.history.MCRMetadataHistoryManager
@@ -117,6 +115,3 @@ MCR.OAIDataProvider.OAI2.ResponseStylesheet=oai/oai2.xsl
 # custom description
 # MCR.OAIDataProvider.OAI.DescriptionURI.EPrints=resource:oai/oai-eprints.xml
 # MCR.OAIDataProvider.OAI.DescriptionURI.Rights=resource:oai/oai-rights.xml
-
-# Relevant classifications for x-path-mapping
-MCR.Category.XPathMapping.ClassIDs=
